refactor(db-orders.js): simplify order retrieval logic by removing try-catch blocks and using async/await directly refactor(validations/orders.js): move order ID input validation logic to a dedicated validation module for better organization
