+++ routes/orders.js
@@ -1,11 +1,9 @@
 // Express promise router
 const Router = require("express-promise-router");
-// HttpError
-const HttpError = require("../utils/HttpError");
-// Joi
-const Joi = require("joi");
 // DB (Knex)
 const { getAllOrders, getOrderById } = require("../db/db-orders");
+// Validations
+const { validateOrderIdInput } = require("../utils/validations/orders");
 
 const ordersRouter = new Router();
 
@@ -18,24 +16,12 @@ ordersRouter.get("/", async (req, res, next) => {
 });
 
 // Gets a specific order
-ordersRouter.get("/:orderId", async (req, res, next) => {
-  // Specify joi schema
-  const schema = Joi.object({
-    order_id: Joi.string().uuid().required(),
-  });
-
-  // Validate the input
-  const { value, error } = schema.validate({
-    order_id: req.params.orderId,
-  });
-
-  // Throw an error if there's an error
-  if (error) {
-    throw new HttpError("Invalid input data", 400);
-  }
-
+ordersRouter.get("/:orderId", validateOrderIdInput, async (req, res, next) => {
   // Query: Get the order
-  const result = await getOrderById(req.user.user_id, value.order_id);
+  const result = await getOrderById(
+    req.user.user_id,
+    req.validatedOrderIdInput.order_id
+  );
 
   // Return the order details
   res.status(200).json(result);
